app/requirements.txt
django
djangorestframework
pymysql
python-dotenv
pandas
requests